refactor : 경로설정 및 경로변경 #102
