+++ src/App.tsx
@@ -4,8 +4,8 @@ import AddressPortal from '@components/AddressPortal';
 import KeywordAddressPortal from '@components/KeywordAddressPortal';
 import LoginPortal from '@components/LoginPortal';
 import NavigationBar from '@components/NavigationBar';
-import OptionPortal from '@components/OptionPortal';
 import SearchPortal from '@components/SearchPortal';
+import OptionPortal from '@components/ShareForm/Option/OptionPortal';
 import useNoticeAlarm from '@hooks/useNoticeAlarm';
 import { routes } from '@router';
 import Styles from '@styles';
