+++ src/components/ShareForm/index.tsx
@@ -1,3 +1,7 @@
-export { default as FileContainer } from '@components/ShareForm/FileContainer';
-export { default as TextContainer } from '@components/ShareForm/TextContainer';
-export { default as ContentDescription } from '@components/ShareForm/ContentDescription';
+export { default as FileContainer } from '@components/ShareForm/Option/FileContainer';
+export { default as TextContainer } from '@components/ShareForm/Option/TextContainer';
+export { default as AddressContainer } from '@components/ShareForm/Option/AddressContainer';
+export { default as DateContainer } from '@components/ShareForm/Option/DateContainer';
+export { default as RecruitmentContainer } from '@components/ShareForm/Option/RecruitmentContainer';
+export { default as OptionPortalButton } from '@components/ShareForm/Option/OptionPortalButton';
+export { default as ContentDescription } from '@components/ShareForm/Option/ContentDescription';
